requirements/development.txt
-r production.txt
pep8
ipython[all]
django-debug-toolbar
django-extensions
#django-pipeline
